🚚structure: 에러 메세지 types -> constants 폴더로 이동
